Moved vuforia init to RobotContainer
